requirements.txt
inflect==5.0.2
ndg-httpsclient==0.4.4
xmltodict
odoo_upgrade @ git+https://github.com/odoo/upgrade-util@master